Remove trailing white space
I get a bunch of noise with git and in my editor with trailing white
space, so I just stripped it out from the source.

This is just a suggestion, but thought it was a nice clean-up
task. If you want to keep the white space just say so and I can figure
out how to get my editor to play nice.