Add support for showing past events
The user can now choose to see "Past Events" and "Upcoming Events" via a
couple of buttons above the events table.

![Preview](https://user-images.githubusercontent.com/16963/113936518-dc947380-97c5-11eb-8c9e-c01d824d587b.gif)

Being able to find past events is very helpful when filling in the event
reimbursement forms and the instructor needs to lookup attendee emails.

This exposes the filter as a query parameter in the events page. That
param is extracted by the `get_events` function when making the request
to WA.

I implemented it this way for a couple of reasons:

1. It centralizes the API request params in one place: the query params.
2. Query params make since here since we are adding parameters to a
   query.
3. I didn't have to change or add any new code paths for loading the
   list of events on the page.