Add migrations
This adds the migrations to get the sqlite database to the current
version in production. If there are existing databases they will need to
be updated.

To update a DB that was created via migrations you can run (same thing
`bin/setup` does):

    ./bin/flask db upgrade

If you see the following error:

    sqlalchemy.exc.OperationalError: (sqlite3.OperationalError) table users already exists

You will need to run:

    ./bin/flask db stamp head

Then rerun the upgrade command above.

This assumes you have a users table already, and it matches the current
migrations defined in migrations/versions/27d0f0a302c2_.py

Why do all this? Allows use to change the schema via code and keep in
sync with the production database so we don't get surprises while
deploying.