Ability to add signoffs by tool
In addition to being able to add signoffs by member we can also pick a
tool and add it to multiple members in one go.

Continued with the jQuery approach, against my better judgment. But it
is nice to limit dependencies.

Updated WAAPI JS to return the result for the callbacks. The promises
were not resolving the values of the API requests, which was less than
helpful.

Added a spinner that is a bit more on-brand.

Favicon!