Link directly to FormBot from Event Details
Thanks to the amazing @Pecacheu and the change in [FormBot][1] we can
now link directly to an Instructor Payment & Maker Sign-off page and
have the event prepopulated. Hopefully this will simplify the process
and reduce confusion and work.

Adds a setting that must be configured in the environment variables for
the base URL. That URL is concatenated with the event ID and a button is
shown in the events page detail view. If the setting does not exist the
button is not rendered.

DEPLOYMENT NOTE: Don't forget to add the FORMBOT_URL setting to the env
when deploying to production.

[1]: https://github.com/Pecacheu/NovaLabs-Instructor-Form/commit/5ac67c9f6ed478011d878bc4af1e64a8e0958201